Move ListPlugins to its own file
